+++ .dockerignore
@@ -1,0 +1,15 @@
+# exclude all by default
+
+*
+
+# exceptions from `exclude all` rule
+
+!package.json
+!yarn.lock
+
+!serviceCallbackFunction/**
+!serviceCallbackRetry/**
+!config/**
+!local.settings.json
+!extensions.csproj
+!host.json
