+++ .gitignore
@@ -62,3 +62,4 @@ node_modules/
 package-lock.json
 .project
 .settings/
+.idea/
