+++ charts/ccpay-functions-nodejs/values.yaml
@@ -1,0 +1,1 @@
+replicas: 1
