+++ deployment.yaml
@@ -1,0 +1,69 @@
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  annotations:
+    deployment.kubernetes.io/revision: "1"
+    kubectl.kubernetes.io/last-applied-configuration: |
+      {"apiVersion":"apps/v1","kind":"Deployment","metadata":{"annotations":{},"labels":{"app":"service"},"name":"service","namespace":"default"},"spec":{"replicas":0,"selector":{"matchLabels":{"app":"service"}},"template":{"metadata":{"labels":{"app":"service"}},"spec":{"containers":[{"env":[{"name":"AzureFunctionsJobHost__functions__0","value":"serviceCallbackFunction"},{"name":"AzureFunctionsJobHost__functions__1","value":"serviceCallbackRetry"}],"envFrom":[{"secretRef":{"name":"service"}}],"image":"hmctssandbox.azurecr.io/service","name":"service"}]}}}}
+  creationTimestamp: "2020-04-16T09:33:06Z"
+  generation: 1
+  labels:
+    app: service
+  name: service
+  namespace: default
+  resourceVersion: "7835936"
+  selfLink: /apis/apps/v1/namespaces/default/deployments/service
+  uid: 551ffbea-2bcc-4a13-8d4c-32f48cc7745d
+spec:
+  progressDeadlineSeconds: 600
+  replicas: 0
+  revisionHistoryLimit: 10
+  selector:
+    matchLabels:
+      app: service
+  strategy:
+    rollingUpdate:
+      maxSurge: 25%
+      maxUnavailable: 25%
+    type: RollingUpdate
+  template:
+    metadata:
+      creationTimestamp: null
+      labels:
+        app: service
+    spec:
+      containers:
+      - env:
+        - name: AzureFunctionsJobHost__functions__0
+          value: serviceCallbackFunction
+        - name: AzureFunctionsJobHost__functions__1
+          value: serviceCallbackRetry
+        envFrom:
+        - secretRef:
+            name: service
+        image: hmctssandbox.azurecr.io/service
+        imagePullPolicy: Always
+        name: service
+        resources: {}
+        terminationMessagePath: /dev/termination-log
+        terminationMessagePolicy: File
+      dnsPolicy: ClusterFirst
+      restartPolicy: Always
+      schedulerName: default-scheduler
+      securityContext: {}
+      terminationGracePeriodSeconds: 30
+status:
+  conditions:
+  - lastTransitionTime: "2020-04-16T09:33:06Z"
+    lastUpdateTime: "2020-04-16T09:33:06Z"
+    message: Deployment has minimum availability.
+    reason: MinimumReplicasAvailable
+    status: "True"
+    type: Available
+  - lastTransitionTime: "2020-04-16T09:33:06Z"
+    lastUpdateTime: "2020-04-16T09:33:06Z"
+    message: ReplicaSet "service-d55775f47" has successfully progressed.
+    reason: NewReplicaSetAvailable
+    status: "True"
+    type: Progressing
+  observedGeneration: 1
