--- functions/.deployment
@@ -1,3 +1,0 @@
-[config]
-SCM_DO_BUILD_DURING_DEPLOYMENT=true
-SCM_COMMAND_IDLE_TIMEOUT=600
