--- functions/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "payment-functions-node",
-  "version": "1.0.0",
-  "private": true,
-  "main": "index.js",
-  "engines": {
-    "node": "8.11.1"
-  },
-  "scripts": {
-    "start": "node index.js",
-    "local": "func host start --debug",
-    "deploy": "func azure functionapp publish ccpay-functions-node",
-    "setup": "func extensions install",
-    "test": "mocha test/*.js"
-  },
-  "dependencies": {
-    "superagent": "^3.8.2",
-    "azure-sb": "latest"
-  },
-  "devDependencies": {
-    "body-parser": "^1.18.3",
-    "chai": "^4.2.0",
-    "debug": "^4.1.0",
-    "express": "^4.16.4",
-    "mocha": "^5.2.0",
-    "proxyquire": "^2.1.0",
-    "sinon": "^5.0.1-next.1",
-    "sinon-chai": "^3.2.0"
-  }
-}
