--- functions/serviceCallbackFunction/function.json
@@ -1,13 +1,0 @@
-{
-  "bindings": [
-    {
-      "topicName": "serviceCallbackTopic",
-      "subscriptionName": "defaultServiceCallbackSubscription",
-      "connection": "ServiceCallbackBusConnection",
-      "name": "serviceCallback",
-      "type": "serviceBusTrigger",
-      "direction": "in"
-    }
-  ],
-  "disabled": false
-}
