--- functions/serviceCallbackRetry/function.json
@@ -1,11 +1,0 @@
-{
-  "bindings": [
-    {
-      "schedule": "%RetrySchedule%",
-      "name": "failedMessagesTimerTrigger",
-      "type": "timerTrigger",
-      "direction": "in"
-    }
-  ],
-  "disabled": false
-}
