+++ host.json
@@ -1,0 +1,38 @@
+{
+  "version": "2.0",
+  "aggregator": {
+    "batchSize": 1000,
+    "flushTimeout": "00:00:30"
+  },
+  "extensions": {
+    "http": {},
+    "queues": {},
+    "serviceBus": {
+      "prefetchCount": 100,
+      "messageHandlerOptions": {
+        "maxConcurrentCalls": 32,
+        "maxAutoRenewDuration": "00:55:00"
+      }
+    }
+  },
+  "healthMonitor": {
+    "enabled": true,
+    "healthCheckInterval": "00:00:10",
+    "healthCheckWindow": "00:02:00",
+    "healthCheckThreshold": 6,
+    "counterThreshold": 0.80
+  },
+
+  "logging": {
+    "fileLoggingMode": "always",
+    "logLevel": {
+      "default": "Trace"
+    },
+    "applicationInsights": {
+      "sampling": {
+        "isEnabled": true,
+        "maxTelemetryItemsPerSecond" : 5
+      }
+    }
+  }
+}
