+++ package.json
@@ -1,0 +1,30 @@
+{
+  "name": "payment-functions-node",
+  "version": "1.0.0",
+  "private": true,
+  "main": "index.js",
+  "engines": {
+    "node": ">=12.16.2"
+  },
+  "scripts": {
+    "start": "node index.js",
+    "local": "func host start",
+    "deploy": "func azure functionapp publish ccpay-functions-node",
+    "setup": "func extensions install",
+    "test": "mocha test/*.js"
+  },
+  "dependencies": {
+    "azure-sb": "^0.11.1",
+    "superagent": "^5.2.2"
+  },
+  "devDependencies": {
+    "body-parser": "^1.19.0",
+    "chai": "^4.2.0",
+    "debug": "^4.1.1",
+    "express": "^4.17.1",
+    "mocha": "^7.1.1",
+    "proxyquire": "^2.1.3",
+    "sinon": "^9.0.2",
+    "sinon-chai": "^3.5.0"
+  }
+}
