+++ serviceCallbackRetry/function.json
@@ -1,0 +1,11 @@
+{
+  "bindings": [
+    {
+      "schedule": "0 */30 * * * *",
+      "name": "failedMessagesTimerTrigger",
+      "type": "timerTrigger",
+      "direction": "in"
+    }
+  ],
+  "disabled": false
+}
